+++ src/main/java/opennote/User/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package opennote.User;
+
+public class UserNotFoundException extends RuntimeException{
+    public UserNotFoundException(Integer id) {
+        super("Could not find user " + id);
+    }
+}
